Fixed vertex attribute.

+++ src/renderer_d3d11.h
@@ -215,7 +215,7 @@ namespace bgfx { namespace d3d11
 		ConstantBuffer* m_constantBuffer;
 
 		PredefinedUniform m_predefined[PredefinedUniform::Count];
-		uint8_t m_attrMask[Attrib::Count];
+		uint16_t m_attrMask[Attrib::Count];
 
 		uint32_t m_hash;
